Displaying menu screen

+++ screen.h
@@ -11,7 +11,7 @@ class Screen
 public:
     Screen(ScreenManager *manager);
     virtual void handleInput(SDL_Event* event) = 0;
-    virtual void update(float dt) = 0;
+    virtual void update(Uint32 dt) = 0;
     virtual void render(SDL_Renderer *ren) = 0;
     virtual void dispose() = 0;
 private:
